Can now evaluate line function and draw the line.

+++ Graphikos/Graphikos/Models/Line.cs
@@ -2,9 +2,9 @@
 {
     public class Line
     {
-        public int X1 { get; set; }
-        public int Y1 { get; set; }
-        public int X2 { get; set; }
-        public int Y2 { get; set; }
+        public double X1 { get; set; }
+        public double Y1 { get; set; }
+        public double X2 { get; set; }
+        public double Y2 { get; set; }
     }
 }
